Move ChipLocationBean to Machine
