+++ SpiNNaker-machine/src/main/java/uk/ac/manchester/spinnaker/machine/bean/MapperFactory.java
@@ -1,0 +1,34 @@
+/*
+ * Copyright (c) 2018 The University of Manchester
+ */
+package uk.ac.manchester.spinnaker.machine.bean;
+
+import static com.fasterxml.jackson.databind.DeserializationFeature.ACCEPT_SINGLE_VALUE_AS_ARRAY;
+import static com.fasterxml.jackson.databind.DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import static com.fasterxml.jackson.databind.PropertyNamingStrategy.SNAKE_CASE;
+import com.fasterxml.jackson.databind.module.SimpleModule;
+
+/**
+ *
+ * @author Christian-B
+ */
+public class MapperFactory {
+    /**
+     * Static method to create the object mapper.
+     *
+     * This method makes sure that all json unmarshallers use the same Mapper
+     *      set up the exact same way.
+     * @return The Object Mapper used by the Spalloc client,
+     */
+    public static ObjectMapper createMapper() {
+        ObjectMapper mapper = new ObjectMapper();
+		SimpleModule module = new SimpleModule();
+		mapper.registerModule(module);
+		mapper.setPropertyNamingStrategy(SNAKE_CASE);
+		mapper.configure(FAIL_ON_UNKNOWN_PROPERTIES, false);
+		mapper.configure(ACCEPT_SINGLE_VALUE_AS_ARRAY, true);
+        return mapper;
+    }
+
+}
